backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
python-dotenv
google-generativeai
groq
firebase-admin
redis
python-multipart
httpx
